[CoreBundle] Add CustomerType
