+++ src/Sylius/Bundle/CoreBundle/Resources/config/validation.xml
@@ -16,6 +16,13 @@
                     xsi:schemaLocation="http://symfony.com/schema/dic/constraint-mapping
                                         http://symfony.com/schema/dic/services/constraint-mapping-1.0.xsd">
 
+    <class name="Sylius\Component\Core\Model\Customer">
+        <constraint name="Sylius\Bundle\UserBundle\Validator\Constraints\RegisteredUser">
+            <option name="message">sylius.customer.email.registered</option>
+            <option name="groups">sylius_customer_guest</option>
+        </constraint>
+    </class>
+
     <class name="Sylius\Component\Core\Model\Channel">
         <property name="taxCalculationStrategy">
             <constraint name="NotBlank" />
