+++ src/Sylius/Bundle/CustomerBundle/Form/Type/CustomerType.php
@@ -21,22 +21,6 @@ use Symfony\Component\OptionsResolver\OptionsResolver;
 class CustomerType extends CustomerProfileType
 {
     /**
-     * @var EventSubscriberInterface
-     */
-    private $addUserFormSubscriber;
-
-    /**
-     * @param string $dataClass
-     * @param string[] $validationGroups
-     * @param EventSubscriberInterface $addUserFormSubscriber
-     */
-    public function __construct($dataClass, array $validationGroups = [], EventSubscriberInterface $addUserFormSubscriber)
-    {
-        parent::__construct($dataClass, $validationGroups);
-        $this->addUserFormSubscriber = $addUserFormSubscriber;
-    }
-
-    /**
      * {@inheritdoc}
      */
     public function buildForm(FormBuilderInterface $builder, array $options)
@@ -57,20 +41,7 @@ class CustomerType extends CustomerProfileType
                 'multiple' => true,
                 'required' => false,
             ])
-            ->addEventSubscriber($this->addUserFormSubscriber)
         ;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function configureOptions(OptionsResolver $resolver)
-    {
-        $resolver->setDefaults([
-            'data_class' => $this->dataClass,
-            'validation_groups' => $this->validationGroups,
-            'cascade_validation' => true,
-        ]);
     }
 
     /**
